feat: Create Streamlit app for Digital Twin Workflow
Refactors the Jupyter notebook into a modular Streamlit application.

The core logic for data loading, pathway analysis, and QUBO-based
drug selection is encapsulated in `core_logic.py`.

The `streamlit_app.py` file provides an interactive web interface
for users to run the analysis with configurable parameters.

Adds a `.gitignore` file to exclude cache, logs, and other
temporary files.
