+++ .gitignore
@@ -1,0 +1,14 @@
+# Ignore the cache directory
+.cache/
+
+# Ignore python cache files
+__pycache__/
+
+# Ignore report files
+*.txt
+
+# Ignore log files
+*.log
+
+# Ignore scratch directory
+jules-scratch/
